+++ walletkit-core/Cargo.toml
@@ -15,7 +15,9 @@ crate-type = ["lib", "staticlib", "cdylib"]
 name = "walletkit_core"
 
 [dependencies]
+hex = "0.4.3"
 ruint = { version = "1.12.3", default-features = false, features = ["alloc"] }
 semaphore = { git = "https://github.com/worldcoin/semaphore-rs", rev = "accb14b", features = ["depth_30"] }
+serde_json = "1.0.133"
 thiserror = "2.0.3"
 uniffi = { workspace = true, features = ["build"] }
