+++ walletkit-core/src/lib.rs
@@ -1,10 +1,8 @@
 #![deny(clippy::all, clippy::pedantic, clippy::nursery)]
 
 pub mod error;
-pub mod field;
 pub mod identity;
 pub mod proof;
-mod utils;
-pub use utils::*;
+pub mod u256;
 
 uniffi::setup_scaffolding!("walletkit_core");
