+++ walletkit-core/src/proof.rs
@@ -7,7 +7,7 @@ pub struct Context {
 
 impl Context {
     #[must_use]
-    pub fn new(app_id: &[u8], action: &[u8]) -> Self {
+    pub fn new(app_id: &[u8], _action: &[u8]) -> Self {
         let external_nullifier = hash_to_field(app_id);
         // TODO: handle action properly
         Self { external_nullifier }
